experiments/dino_v3/requirements.txt
torch
torchvision
transformers
huggingface_hub
pillow
numpy
scipy
scikit-learn
safetensors
